# CASMCMS-9211: Improve performance of configuration delete operation