When a CFS configuration is deleted, it checks to make sure it is not in use before deleting it. The method it uses to do this is to look at every component and make a list of every configuration that is in use, and then see if this particular configuration is in that list. This has the virtue of reusing code, but it can take a long time. On a large scale system like mug, the delete operation took over 35 seconds.

This PR improves that procedure in two ways:
1. Instead of listing all components, it queries specifically for components that have that particular configuration as their desired configuration. This means that the filtering for relevant components happens much closer to the database calls.
2. When doing that query, set a limit of 1. That means that CFS will respond to us after it finds a single matching component. We only care whether any such components exist and have no interest in the full list, so this is better for us.

I tested this on mug. After making these changes, deleting an unused configuration took 13 seconds, and attempting to delete an in-use configuration took 2 seconds (previously even this case took 35 seconds).